Upgrade to JasperFx/Marten/Weasel rc; Saga.Version is permanently int (#2872)
Bumps the critter-stack dependencies to their rc lines (holding Polecat,
which has no rc yet):
  - JasperFx family       -> 2.0.0-rc.2 / 5.0.0-rc.2
  - Marten / AspNetCore / Newtonsoft -> 9.0.0-rc.1
  - Weasel family         -> 9.0.0-rc.1
  - Polecat               -> held at 4.0.0-alpha.10

JasperFx 2.0 rc split numeric versioning into two interfaces:
  - IRevisioned.Version   : int   (optimistic-concurrency revision)
  - ILongVersioned.Version: long  (event-stream version)

Accordingly:
  - Saga.Version reverts from long back to int (permanent design decision)
    so sagas implement IRevisioned directly without a shadow override.
  - MartenIntegration resolves the saga revision member via
    GetProperty(nameof(Saga.Version), typeof(int)) to match.
  - LoadDocumentFrame keeps the generated expectedSagaRevision as a long
    because IDocumentOperations.UpdateRevision(entity, long) takes a long;
    the int Saga.Version + 1 widens into it. Comment corrected.
  - Event-sourced aggregate test types that track a long stream version
    (SideEffects1, GpStreamAggregate, GSimpleAggregate) now implement
    ILongVersioned instead of IRevisioned.

Test/sample compile fixes for rc namespace relocations:
  - IRevisioned: Marten.Metadata -> JasperFx
  - TenancyStyle: Marten.Storage -> JasperFx.MultiTenancy
  - Identity/IdentityAttribute: Marten.Schema -> JasperFx
  - ISoftDeleted: Marten.Metadata -> JasperFx.Metadata
  - PolecatTests: alias [Identity] to Wolverine.Polecat.IdentityAttribute
    to disambiguate from the lifted JasperFx.IdentityAttribute

Full wolverine.slnx builds clean (0 warnings, 0 errors) on net9.0/net10.0.

+++ src/Persistence/Wolverine.Marten/MartenIntegration.cs
@@ -175,19 +175,15 @@ internal class MartenOverrides : IConfigureMarten
             {
                 mapping.UseNumericRevisions = true;
                 // GetProperty(name, returnType) — not GetProperty(name) — because
-                // saga subclasses in the wild (pre-6.0) sometimes declare
-                // `public new int Version` to work around the 5.x mismatch
-                // between Saga.Version: int and Marten's IRevisioned.Version:
-                // long. After the 6.0 widening of Saga.Version to long
-                // (matching Marten 9.0.0-alpha.2's IRevisioned.Version), the
-                // single-arg overload throws AmbiguousMatchException on those
-                // subclasses because the inherited long and the derived int
-                // are two distinct properties. Filtering by return type picks
-                // the inherited long property unambiguously and treats the
-                // derived shadow as harmless dead code (which it now is, per
-                // the Saga.Version XML doc).
+                // saga subclasses in the wild sometimes declare a shadowing
+                // `public new ... Version` property. Saga.Version is permanently
+                // an int (it aligns with JasperFx 2.0 rc's IRevisioned.Version,
+                // which is an int; the long-versioned event-sourcing case is a
+                // separate ILongVersioned.Version). Filtering by the int return
+                // type picks the canonical Saga.Version revision property and
+                // ignores any derived shadow of a different type.
                 mapping.Metadata.Revision.Member = mapping.DocumentType.GetProperty(
-                    nameof(Saga.Version), typeof(long))!;
+                    nameof(Saga.Version), typeof(int))!;
             }
         });
     }
